refactor: split cashflow planning dtos
